tests/wpt/web-platform-tests/app-history/navigate/navigate-detach.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = resolve);
  let iframe_exception = i.contentWindow.DOMException;
  i.contentWindow.appHistory.onnavigate = () => i.remove();
  await promise_rejects_dom(t, 'AbortError', iframe_exception, i.contentWindow.appHistory.navigate("#"));
}, "Detach inside onnavigate");
</script>
